refactor: fix module name
