+++ Cargo.toml
@@ -8,3 +8,4 @@ edition = "2021"
 [dependencies]
 rusqlite = { version = "0.27.0", features = ["bundled"] }
 clap = { version = "3.1.6", features = ["derive"] }
+tabwriter = "1"
