+++ src/presentation/mod.rs
@@ -1,1 +1,1 @@
-pub mod commands;
+pub mod command;
